Unify multifile test architecture under shared AsyncMultiFileFixture
- Replace environment variable logging with constexpr approach
- Extract common file handling into reusable FileTestFixture
- Remove ~160 lines of duplicate fixture code across test files
- Fix clang-tidy warnings for coroutine parameters and missing destructor
- Update BUILD.bazel dependencies for proper fixture sharing
- All tests pass (14/14) with improved code organization
